requirements.txt
flask
gunicorn
azure-ai-projects
azure-identity
azure-ai-agents
python-dotenv